% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcast_data_split.R
\name{annual_reports_animal_hosts_training}
\alias{annual_reports_animal_hosts_training}
\title{Get cases training set (~80\%)}
\usage{
annual_reports_animal_hosts_training(conn)
}
\value{
a tibble
}
\description{
Get cases training set (~80\%)
}
