% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_augment_lookup.R
\name{baseline_augment_lookup}
\alias{baseline_augment_lookup}
\title{Support function to generate the augment lookup for the baseline model}
\usage{
baseline_augment_lookup(conn)
}
\arguments{
\item{conn}{either repel_local_conn or repel_remote_conn objects from 'repeldata'}
}
\description{
Support function to generate the augment lookup for the baseline model
}
