% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_augment_lookup.R
\name{baseline_augment_lookup_specs}
\alias{baseline_augment_lookup_specs}
\title{column specs for reading in baseline_augment_lookup csv}
\usage{
baseline_augment_lookup_specs()
}
\description{
column specs for reading in baseline_augment_lookup csv
}
