% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_lag_lookup.R
\name{case_lag_lookup}
\alias{case_lag_lookup}
\title{Support function to generate lag lookup for case data}
\usage{
case_lag_lookup(conn)
}
\arguments{
\item{conn}{either repel_local_conn or repel_remote_conn objects from 'repeldata'}
}
\description{
Support function to generate lag lookup for case data
}
