% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_lag_lookup.R
\name{case_lag_lookup_specs}
\alias{case_lag_lookup_specs}
\title{column specs for reading in case_lag_lookup csv}
\usage{
case_lag_lookup_specs()
}
\description{
column specs for reading in case_lag_lookup csv
}
