% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_augment_lookup.R
\name{get_baseline_augment_lookup}
\alias{get_baseline_augment_lookup}
\title{read in augment lookup for the baseline model}
\usage{
get_baseline_augment_lookup()
}
\description{
read in augment lookup for the baseline model
}
