% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cached_network.R
\name{get_cached_network_predictions}
\alias{get_cached_network_predictions}
\title{Get cached network model augmented data and model predictions}
\usage{
get_cached_network_predictions(conn)
}
\arguments{
\item{conn}{repeldata connection}
}
\value{
dataframe containing full dataset and predictions
}
\description{
Get cached network model augmented data and model predictions
}
