% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cached_network.R
\name{get_cached_network_random_effects}
\alias{get_cached_network_random_effects}
\title{Get cached model random effect coefficients}
\usage{
get_cached_network_random_effects()
}
\value{
dataframe with coefficient by disease and variable
}
\description{
Get cached model random effect coefficients
}
