% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_disease_status_predictions.R
\name{get_disease_status_predict}
\alias{get_disease_status_predict}
\title{Get disease status and predicted outbreak prob by country}
\usage{
get_disease_status_predict(
  network_boost_augment_predict,
  network_lme_augment_predict,
  country_iso3c,
  diseases = get_oie_high_importance_diseases()
)
}
\arguments{
\item{network_boost_augment_predict}{output of \code{get_cached_nowcast_predictions()}}

\item{network_lme_augment_predict}{output of \code{get_cached_network_predictions()}}

\item{country_iso3c}{iso3c(s) of countries}

\item{disease}{disease(s) name(s)}

\item{month}{in format of "yyyy-mm-01". if NULL, uses latest month in dataset}
}
\value{
dataframe containing actual values ('x'), coefficients ('coef') and variable importance
}
\description{
Get disease status and predicted outbreak prob by country
}
