% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lag.R
\name{get_lag}
\alias{get_lag}
\title{Get  lags in cases, disease status, and control measures for nowcast model}
\usage{
get_lag(six_month_processed, lags = 3)
}
\description{
Get  lags in cases, disease status, and control measures for nowcast model
}
