% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_network_variable_importance.R
\name{get_network_origin_contribution_import_risk}
\alias{get_network_origin_contribution_import_risk}
\title{Get contribution of origin countries to import risk}
\usage{
get_network_origin_contribution_import_risk(
  conn,
  country_iso3c,
  diseases = get_oie_high_importance_diseases(),
  month = NULL
)
}
\arguments{
\item{conn}{repeldata connection}

\item{country_iso3c}{iso3c(s) of countries}

\item{month}{in format of "yyyy-mm-01". if NULL, uses latest month}

\item{disease}{disease(s) name(s)}
}
\value{
dataframe containing sum of variable importance by country origin
}
\description{
Get contribution of origin countries to import risk
}
