% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_network_variable_importance.R
\name{get_network_variable_importance_with_origins}
\alias{get_network_variable_importance_with_origins}
\title{Get variable importance for given disease and month, by country}
\usage{
get_network_variable_importance_with_origins(
  disagg_network_augment_predict,
  randef,
  country_iso3c,
  diseases = get_oie_high_importance_diseases(),
  month = NULL
)
}
\arguments{
\item{disagg_network_augment_predict}{output of \code{get_cached_disagg_network_predictions()}}

\item{randef}{output of \code{get_cached_network_random_effects()}}

\item{country_iso3c}{iso3c(s) of countries}

\item{month}{in format of "yyyy-mm-01". if NULL, uses latest month in dataset}

\item{disease}{disease(s) name(s)}
}
\value{
dataframe containing actual values ('x'), coefficients ('coef') and variable importance
}
\description{
Get variable importance for given disease and month, by country
}
