% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_objects.R
\name{impact_gam_model}
\alias{impact_gam_model}
\title{Produces impact model object}
\usage{
impact_gam_model(impact_model = NULL)
}
\value{
a list with description and classes
}
\description{
Produces impact model object
}
