% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_objects.R
\name{network_brms_model}
\alias{network_brms_model}
\title{Produces network brms model object}
\usage{
network_brms_model(network_model = NULL, network_scaling_values = NULL)
}
\value{
a list with description and classes
}
\description{
Produces network brms model object
}
