% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_objects.R
\name{network_impact_model}
\alias{network_impact_model}
\title{Produces impact model object}
\usage{
network_impact_model(impact_model = NULL)
}
\value{
a list with description and classes
}
\description{
Produces impact model object
}
