% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_objects.R
\name{network_lme_model}
\alias{network_lme_model}
\title{Produces network lme model object}
\usage{
network_lme_model(network_model = NULL)
}
\value{
a list with description and classes
}
\description{
Produces network lme model object
}
