% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcast_models.R
\name{nowcast_bart_model}
\alias{nowcast_bart_model}
\title{Produces nowcast bart model object}
\usage{
nowcast_bart_model()
}
\value{
a list with description and classes
}
\description{
Produces nowcast bart model object
}
