% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcast.R
\name{nowcast_baseline_augment}
\alias{nowcast_baseline_augment}
\title{Augment data with predictions}
\usage{
nowcast_baseline_augment(conn, newdata, n_semesters = NULL)
}
\value{
a tibble
}
\description{
Augment data with predictions
}
