% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcast.R
\name{nowcast_baseline_model}
\alias{nowcast_baseline_model}
\title{Get baseline cases nowcast}
\usage{
nowcast_baseline_model(conn, dat)
}
\value{
a list with predictions and metadata
}
\description{
Get baseline cases nowcast
}
