% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcast_models.R
\name{nowcast_boost_model}
\alias{nowcast_boost_model}
\title{Produces nowcast xgboost model object}
\usage{
nowcast_boost_model()
}
\value{
a list with description and classes
}
\description{
Produces nowcast xgboost model object
}
