% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_objects.R
\name{nowcast_gam_model}
\alias{nowcast_gam_model}
\title{Produces nowcast gam model object}
\usage{
nowcast_gam_model()
}
\value{
a list with description and classes
}
\description{
Produces nowcast gam model object
}
