% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepvar.R
\name{prepvar}
\alias{prepvar}
\title{Prepare a variable for fitting transformation and dealing with NA and zero}
\usage{
prepvar(x, na_val = "auto", zero_val = "zero", trans_fn = identity)
}
\arguments{
\item{x}{the variable to transform}

\item{na_val}{the value to convert NAs to
"auto" will select a value well below the range of the rest}

\item{zero_val}{the value to conver zeros to.
"zero" treat zeros as any other value. "auto" will select a
value well below the range of the rest}

\item{trans_fn}{function to transform other values with, e.g., \code{log10}}
}
\value{
A vector of length x
}
\description{
Prepare a variable for fitting transformation and dealing with NA and zero
}
\examples{
set.seed(0)
z <- round(abs(rnorm(1000, sd = 4)))
z[sample(1:1000, 100)] <- NA
p <- prepvar(z, na_val = "auto", zero_val = "auto", trans_fn = log)
hist(p)
}
