% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{repel_augment.impact_model}
\alias{repel_augment.impact_model}
\title{Augment nowcast baseline model object}
\usage{
\method{repel_augment}{impact_model}(model_object, conn, newdata)
}
\description{
Augment nowcast baseline model object
}
