% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{repel_augment.network_baseline}
\alias{repel_augment.network_baseline}
\title{Augment baseline model object}
\usage{
\method{repel_augment}{network_baseline}(model_object, conn, newdata)
}
\arguments{
\item{model_object}{network_lme object}

\item{conn}{connection to repel db}

\item{newdata}{dataframe that has been preprocessed through \code{repel_init()} and optionally split into training/validation. contains country, disease, month, and outbreak status.}
}
\description{
Augment baseline model object
}
