% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{repel_augment.network_lme}
\alias{repel_augment.network_lme}
\title{Augment nowcast baseline model object}
\usage{
\method{repel_augment}{network_lme}(model_object, conn, newdata)
}
\description{
Augment nowcast baseline model object
}
