% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{repel_augment.network_model}
\alias{repel_augment.network_model}
\title{Augment nowcast baseline model object}
\usage{
\method{repel_augment}{network_model}(model_object, conn, newdata, sum_country_imports = TRUE)
}
\description{
Augment nowcast baseline model object
}
