% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{repel_augment.nowcast_baseline}
\alias{repel_augment.nowcast_baseline}
\title{Augment nowcast baseline model object}
\usage{
\method{repel_augment}{nowcast_baseline}(model_object, conn, subset = NULL)
}
\arguments{
\item{model_object}{nowcast model object}

\item{conn}{connection to repel db}

\item{subset}{optional dataframe that contains country_iso3c, report_year, report_semester, disease, disease_population, taxa. Used to subset the output (e.g., for validation/training)}
}
\description{
Augment nowcast baseline model object
}
