% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{repel_augment.nowcast_gam}
\alias{repel_augment.nowcast_gam}
\title{Augment nowcast GAM model object}
\usage{
\method{repel_augment}{nowcast_gam}(model_object, conn, newdata, rare = 1000)
}
\description{
Augment nowcast GAM model object
}
