% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcast_augment.R
\name{repel_augment.nowcast_tree}
\alias{repel_augment.nowcast_tree}
\title{Augment nowcast bart model object}
\usage{
\method{repel_augment}{nowcast_tree}(model_object, conn, newdata)
}
\description{
Augment nowcast bart model object
}
