% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcast_data_split.R
\name{repel_cases_validate}
\alias{repel_cases_validate}
\title{Hold out validation set (~20\%)}
\usage{
repel_cases_validate(conn)
}
\value{
a tibble
}
\description{
Hold out validation set (~20\%)
}
