% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_disease_names.R
\name{repel_clean_disease_names.network_model}
\alias{repel_clean_disease_names.network_model}
\title{Clean network disease names}
\usage{
\method{repel_clean_disease_names}{network_model}(model_object, df)
}
\arguments{
\item{model_object}{network model object}

\item{df}{dataframe with disease column}
}
\description{
Clean network disease names
}
