% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_disease_names.R
\name{repel_clean_disease_names.repel_model}
\alias{repel_clean_disease_names.repel_model}
\title{Clean disease names}
\usage{
\method{repel_clean_disease_names}{repel_model}(model_object, df)
}
\arguments{
\item{model_object}{model object (network or nowcast)}

\item{df}{dataframe with disease column}
}
\description{
Clean disease names
}
