% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{repel_fit.network_baseline}
\alias{repel_fit.network_baseline}
\title{Fit network baseline model object}
\usage{
\method{repel_fit}{network_baseline}(
  model_object,
  augmented_data,
  predictor_vars,
  verbose = interactive()
)
}
\description{
Fit network baseline model object
}
