% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{repel_fit.network_brms}
\alias{repel_fit.network_brms}
\title{Fit network BRMS model object}
\usage{
\method{repel_fit}{network_brms}(
  model_object,
  augmented_data,
  predictor_vars,
  output_directory,
  verbose = interactive()
)
}
\value{
list containing predicted count and whether disease is expected or not (T/F)
}
\description{
Fit network BRMS model object
}
