% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcast_fit.R
\name{repel_fit.nowcast_bart}
\alias{repel_fit.nowcast_bart}
\title{Fit nowcast BART model object}
\usage{
\method{repel_fit}{nowcast_bart}(
  model_object,
  augmented_data,
  output_directory,
  n_chains = dbarts::guessNumCores(),
  verbose = interactive()
)
}
\value{
list containing predicted count and whether disease is expected or not (T/F)
}
\description{
Fit nowcast BART model object
}
