% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcast_fit.R
\name{repel_fit.nowcast_boost}
\alias{repel_fit.nowcast_boost}
\title{Fit nowcast Boost model object}
\usage{
\method{repel_fit}{nowcast_boost}(
  model_object,
  augmented_data,
  output_directory,
  verbose = interactive()
)
}
\value{
list containing predicted count and whether disease is expected or not (T/F)
}
\description{
Fit nowcast Boost model object
}
