% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{repel_forecast.network_lme}
\alias{repel_forecast.network_lme}
\title{forecast network lme model object}
\usage{
\method{repel_forecast}{network_lme}(model_object, conn, newdata, use_cache = TRUE)
}
\arguments{
\item{model_object}{nowcast model object}

\item{conn}{connection to repel db}

\item{newdata}{\itemize{
\item dataframe or tibble with fields: country_iso3c, report_year, report_semester, disease, disease_population, taxa
}}

\item{use_cache}{whether to use cached prediced data (TRUE/FALSE)}
}
\description{
forecast network lme model object
}
