% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{repel_forecast.network_model}
\alias{repel_forecast.network_model}
\title{forecast nowcast network model object}
\usage{
\method{repel_forecast}{network_model}(model_object, conn, newdata)
}
\arguments{
\item{model_object}{network model object}

\item{conn}{connection to repel db}

\item{newdata}{data to augment/predict}
}
\description{
forecast nowcast network model object
}
