% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{repel_forecast.nowcast_baseline}
\alias{repel_forecast.nowcast_baseline}
\title{forecast nowcast baseline model object}
\usage{
\method{repel_forecast}{nowcast_baseline}(model_object, conn, newdata)
}
\arguments{
\item{model_object}{nowcast model object}

\item{conn}{connection to repel db}
}
\description{
forecast nowcast baseline model object
}
