% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcast_forecast.R
\name{repel_forecast.nowcast_boost}
\alias{repel_forecast.nowcast_boost}
\title{forecast nowcast boost model object}
\usage{
\method{repel_forecast}{nowcast_boost}(model_object, conn, newdata, use_cache = TRUE)
}
\description{
forecast nowcast boost model object
}
