% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{repel_forecast.nowcast_model}
\alias{repel_forecast.nowcast_model}
\title{forecast nowcast model object}
\usage{
\method{repel_forecast}{nowcast_model}(model_object, conn, subset = NULL, six_month_processed = NULL)
}
\arguments{
\item{model_object}{nowcast model object}

\item{conn}{connection to repel db}

\item{subset}{optional dataframe that contains country_iso3c, report_year, report_semester, disease, disease_population, taxa. Used to subset the output (e.g., for validation/training)}

\item{six_month_processed}{optional dataframe that has been processed by repel_init(). If not provided, it will be generated here.}
}
\description{
forecast nowcast model object
}
