% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{repel_forecast.repel_model}
\alias{repel_forecast.repel_model}
\title{forecast nowcast baseline model object}
\usage{
\method{repel_forecast}{repel_model}(model_object, conn, newdata)
}
\arguments{
\item{model_object}{nowcast model object}

\item{conn}{connection to repel db}

\item{newdata}{data to augment/predict}
}
\description{
forecast nowcast baseline model object
}
