% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{repel_init.network_model}
\alias{repel_init.network_model}
\title{Preprocess network data}
\usage{
\method{repel_init}{network_model}(
  model_object,
  conn,
  outbreak_reports_events = NULL,
  remove_single_country_disease = TRUE,
  remove_non_primary_taxa_disease = TRUE,
  clean_disease_names = TRUE
)
}
\arguments{
\item{model_object}{network model object}

\item{conn}{connection to repel db}

\item{outbreak_reports_events}{optional to provide outbreak_reports_events dataframe. This is used when providing new data. Default is to read full outbreak_reports_events from conn for model fitting.}

\item{remove_single_country_disease}{whether to remove diseases that occur in only one country. Default is TRUE for model fitting. Use FALSE for new data.}

\item{remove_non_primary_taxa_disease}{whether to remove diseases that do not occur in previously identified taxa for that disease. Default is TRUE for model fitting. Use FALSE for new data.}

\item{clean_disease_names}{defaults to TRUE but needs to be false when running disease name cleaning lookup}
}
\description{
Preprocess network data
}
