% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lag.R
\name{repel_lag.nowcast_model}
\alias{repel_lag.nowcast_model}
\title{Get case and status lags}
\usage{
\method{repel_lag}{nowcast_model}(
  model_object,
  conn,
  six_month_reports_summary = NULL,
  newdata,
  lags = 1:3,
  control_measures = FALSE
)
}
\description{
Get case and status lags
}
