% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{repel_predict.network_brms}
\alias{repel_predict.network_brms}
\title{Predict from network brms model object}
\usage{
\method{repel_predict}{network_brms}(model_object, newdata)
}
\value{
vector containing predicted probability of an outbreak
}
\description{
Predict from network brms model object
}
