% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{repel_predict.network_lme}
\alias{repel_predict.network_lme}
\title{Predict from network lme model object}
\usage{
\method{repel_predict}{network_lme}(model_object, newdata)
}
\value{
vector containing predicted probability of an outbreak
}
\description{
Predict from network lme model object
}
