% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcast_predict.R
\name{repel_predict.nowcast_bart}
\alias{repel_predict.nowcast_bart}
\title{Predict from nowcast BART model object}
\usage{
\method{repel_predict}{nowcast_bart}(model_object, augmented_data)
}
\value{
list containing predicted count and whether disease is expected or not (T/F)
}
\description{
Predict from nowcast BART model object
}
