% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{repel_predict.nowcast_baseline}
\alias{repel_predict.nowcast_baseline}
\title{Predict from nowcast baseline model object}
\usage{
\method{repel_predict}{nowcast_baseline}(model_object, newdata)
}
\value{
list containing predicted count and whether disease is expected or not (T/F)
}
\description{
Predict from nowcast baseline model object
}
