% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_split.R
\name{repel_split.network_model}
\alias{repel_split.network_model}
\title{Lookup train/val split for network model data}
\usage{
\method{repel_split}{network_model}(model_object, conn)
}
\arguments{
\item{model_object}{network model object}

\item{conn}{connection to repel db}
}
\description{
Lookup train/val split for network model data
}
