% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcast_data_split.R
\name{repel_training.nowcast_model}
\alias{repel_training.nowcast_model}
\title{Get cases training set (~80\%)}
\usage{
\method{repel_training}{nowcast_model}(model_object, conn)
}
\value{
a tibble
}
\description{
Get cases training set (~80\%)
}
