% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_split.R
\name{repel_training.repel_model}
\alias{repel_training.repel_model}
\title{Get cases training set (~80\%)}
\usage{
\method{repel_training}{repel_model}(model_object, conn)
}
\arguments{
\item{model_object}{network model object}

\item{conn}{connection to repel db}
}
\value{
a tibble
}
\description{
Get cases training set (~80\%)
}
