% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcast_data_split.R
\name{repel_validation.nowcast_model}
\alias{repel_validation.nowcast_model}
\title{Hold out validation set (~20\%)}
\usage{
\method{repel_validation}{nowcast_model}(model_object, conn)
}
\value{
a tibble
}
\description{
Hold out validation set (~20\%)
}
