% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repelpredict-package.R
\docType{package}
\name{repelpredict-package}
\alias{repelpredict}
\alias{repelpredict-package}
\title{repelpredict: Modeling veterinary disease occurrence and spread using repel database}
\description{
This package contains functions that pull data from the repel database, perform data processing and augmentation, fit models, and run predictions from models.
}
\author{
\strong{Maintainer}: Ross Noam \email{noam.ross@gmail.com} (\href{https://orcid.org/0000-0002-2136-0000}{ORCID})

}
\keyword{internal}
