% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_data_split.R, R/nowcast_data_split.R
\name{annual_reports_animal_hosts_validation}
\alias{annual_reports_animal_hosts_validation}
\title{Hold out validation set (~20\%)}
\usage{
annual_reports_animal_hosts_validation(conn)

annual_reports_animal_hosts_validation(conn)
}
\value{
a tibble

a tibble
}
\description{
Hold out validation set (~20\%)

Hold out validation set (~20\%)
}
