% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_disease_status_predict.R
\name{get_disease_status_predict}
\alias{get_disease_status_predict}
\title{Get disease status and predicted outbreak prob by country}
\usage{
get_disease_status_predict(
  conn,
  country_iso3c,
  diseases = get_oie_high_importance_diseases()
)
}
\arguments{
\item{conn}{repeldata connection}

\item{country_iso3c}{iso3c(s) of countries}

\item{disease}{disease(s) name(s)}
}
\value{
dataframe containing disease status (reported/predicted), reported endemic/outbreak status, predicted imported outbreak prob
}
\description{
Get disease status and predicted outbreak prob by country
}
