% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_objects.R
\name{network_baseline_model}
\alias{network_baseline_model}
\title{Produces network baseline model object}
\usage{
network_baseline_model()
}
\value{
a list with description and classes
}
\description{
Produces network baseline model object
}
