% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_objects.R
\name{nowcast_baseline_model}
\alias{nowcast_baseline_model}
\title{Produces nowcast baseline model object}
\usage{
nowcast_baseline_model()
}
\value{
a list with description and classes
}
\description{
Produces nowcast baseline model object
}
