% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcast.R
\name{nowcast_baseline_predict}
\alias{nowcast_baseline_predict}
\title{Get baseline cases nowcast}
\usage{
nowcast_baseline_predict(conn, country_iso3c, year, disease, taxa)
}
\description{
Get baseline cases nowcast
}
