% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{repel_augment.network_lme}
\alias{repel_augment.network_lme}
\title{Augment network model object}
\usage{
\method{repel_augment}{network_lme}(model_object, conn, newdata, sum_country_imports = TRUE)
}
\arguments{
\item{model_object}{network_lme object}

\item{conn}{connection to repel db}

\item{newdata}{dataframe that has been preprocessed through \code{repel_init()} and optionally split into training/validation. contains country, disease, month, and outbreak status.}

\item{sum_country_imports}{whether or not to sum all imports. FALSE results in data disaggregated by origin countries. Default is TRUE.}
}
\description{
Augment network model object
}
