% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{repel_augment.network_model}
\alias{repel_augment.network_model}
\title{Augment steps for baseline and full model}
\usage{
\method{repel_augment}{network_model}(model_object, conn, newdata, sum_country_imports = TRUE)
}
\arguments{
\item{model_object}{network model object}

\item{conn}{connection to repel db}

\item{newdata}{dataframe that has been preprocessed through \code{repel_init()} and optionally split into training/validation. contains country, disease, month, and outbreak status.}
}
\description{
Augment steps for baseline and full model
}
