% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{repel_augment.nowcast_baseline}
\alias{repel_augment.nowcast_baseline}
\title{Augment nowcast baseline model object}
\usage{
\method{repel_augment}{nowcast_baseline}(model_object, conn, newdata)
}
\description{
Augment nowcast baseline model object
}
