% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{repel_augment.nowcast_boost}
\alias{repel_augment.nowcast_boost}
\title{Augment nowcast bart model object}
\usage{
\method{repel_augment}{nowcast_boost}(model_object, conn, newdata)
}
\description{
Augment nowcast bart model object
}
