% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{repel_fit.network_model}
\alias{repel_fit.network_model}
\title{Feature engineering (pre-fit) steps for baseline and full model}
\usage{
\method{repel_fit}{network_model}(model_object, augmented_data, predictor_vars)
}
\description{
Feature engineering (pre-fit) steps for baseline and full model
}
