% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{repel_predict.nowcast_boost}
\alias{repel_predict.nowcast_boost}
\title{Predict from nowcast xgboost model object}
\usage{
\method{repel_predict}{nowcast_boost}(model_object, newdata, use_cache = TRUE)

\method{repel_predict}{nowcast_boost}(model_object, newdata, use_cache = TRUE)
}
\value{
list containing predicted count and whether disease is expected or not (T/F)

list containing predicted count and whether disease is expected or not (T/F)
}
\description{
Predict from nowcast xgboost model object

Predict from netwrok lme model object
}
