% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{repel_predict.nowcast_boost}
\alias{repel_predict.nowcast_boost}
\title{Predict from nowcast xgboost model object}
\usage{
\method{repel_predict}{nowcast_boost}(model_object, newdata, use_cache = TRUE)
}
\value{
vector containing predicted count
}
\description{
Predict from nowcast xgboost model object
}
