% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repel-predict-package.R
\docType{package}
\name{repelpredict-package}
\alias{repelpredict}
\alias{repelpredict-package}
\title{repelpredict: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Ross Noam \email{noam.ross@gmail.com} (0000-0002-2136-0000)

}
\keyword{internal}
